backend/requirements.txt
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==1.7.2
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.7.1
Django==1.11.13
django-cors-headers==2.2.0
django-filter==1.1.0
django-mysql==2.2.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
enum34==1.1.6
et-xmlfile==1.0.1
idna==2.6
ipaddress==1.0.22
jdcal==1.4
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
MySQL-python==1.2.5
numpy==1.14.3
openpyxl==2.5.3
pathlib==1.0.1
plac==0.9.6
preshed==1.0.0
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.6.1
pytz==2018.4
regex==2017.4.5
requests==2.18.4
six==1.11.0
spacy==2.0.11
termcolor==1.1.0
thinc==6.10.2
toolz==0.9.0
tqdm==4.23.3
typing==3.5.3.0
ujson==1.35
urllib3==1.22
wrapt==1.10.11
